requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
django-jalali-date==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exponent-server-sdk==2.0.0
idna==3.6
jdatetime==4.1.1
Pillow==10.0.0
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.1.0
